requirements.txt
aniso8601==9.0.1
attrs==21.4.0
click==8.0.4
Flask==2.0.3
flask-restx==0.5.1
importlib-metadata==4.11.2
itsdangerous==2.1.0
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.1.0
pyrsistent==0.18.1
pytz==2021.3
six==1.16.0
typing-extensions==4.1.1
Werkzeug==2.0.3
flask_sqlalchemy==2.5.1
SQLAlchemy==1.4.31
zipp==3.7.0
marshmallow==3.14.1
pyjwt==2.3.0
pytest~=7.0.1